Define equality between faces and Rubik's cube
This commit defines what it means for two faces and moreover two Rubik's
cubes to be equal. Two Rubik's cubes are said to be equal if they have
equal faces, and two faces are said to be equal if they contain the same
pieces. This is reflected in the newly added unit tests.