Manipulate dummy user input
This commit manipulates a string which is intended to represent the
input of the state of the Rubik's cube from the user. Using this input
string, the faces and the pieces can be deciphered, allowing the state
of the cube to be determined by the user in the future. This
functionality is reflected in the newly added unit tests.