Pass pieces to face as parameter
This commit removes the hardcoding of the pieces of a face based on what
face they are and moves toward a solution where the state of the face is
determined by the arguments passed to the constructor. The Face
constructor now takes the name of the face (e.g. F (Face), U (Up), R
(Right), B (Back), L (Left), D (Down)) along with the state of the face
as a string (e.g. "RRRRRRRRR" (a face of nine Red pieces). The unit
tests reflect this change.