Pass faces to Rubik's cube as parameter
This commit enables the state of the Rubik's cube to be passed as an
argument to the constructor as opposed to hardcoding a single state of
every Rubik's cube as was previously done. The constructor can be passed
a string which defines the faces of the Rubik's cube e.g. "WWWWWWWWW
YYYYYYYYY RRRRRRRRR OOOOOOOOO BBBBBBBBB GGGGGGGGG" represents a solved
Rubik's cube. The faces are passed in the order Face (F), Back (B),
Right (R), Left (L), Up (U), Down (D). The pieces are given in the
format where the first three letters represent the top layer, the second
three letters represent the middle layer and the final three characters
represent the bottom layer - all read from left to right. The unit tests
test this functionality.