Validate parameters of Rubik's cube
This commit verifies that the argument passed as a parameter to the
Rubik's cube is either "SOLVED", "RANDOM" or a string denoting the
makeup of all six faces. The latter still needs to be properly
implemented and is hardcoded as of now. Unit tests have been added to
reflect the current functionality.