Extracted validation of Rubik's cube from class
This commit removes the methods from the RubiksCube class which were
previously verifying the validity of the cube. The intention in doing
this is to ensure that the input passed to the RubiksCube constructor is
in fact valid and therefore it is only possible to create valid cubes.
The validity check does not cover all edge cases as of yet, but this
adjustment enables a greater number of user inputs of cube
configurations to be verified as valid. The unit tests reflect this
functionality.