Add main to RubiksCubeSolver
This commit introduces a main function to the Rubik's cube solver,
allowing the user to interact with the program. This initial program
when run, prompts the user to enter the state of the Rubik's cube. If a
valid state is provided, the specified Rubik's cube object is created
and printed to the console. __repr__ methods have been added to the
RubiksCube and Face classes to enable printing of the cube. These
methods still require unit tests.